+++ pallets/common/Cargo.toml
@@ -10,6 +10,7 @@ polymesh-primitives-derive = { path = "../../primitives_derive", default-feature
 
 # Our Pallets
 pallet-base = { workspace = true, default-features = false  }
+pallet-permissions = { workspace = true, default-features = false  }
 
 # Other
 serde = { version = "1.0.112", default-features = false }
@@ -54,6 +55,7 @@ std = [
 	"frame-system/std",
 	"pallet-timestamp/std",
 	"pallet-base/std",
+	"pallet-permissions/std",
 	"polymesh-primitives/std",
 	"serde/std",
 	"serde_derive",
@@ -66,5 +68,6 @@ std = [
 ]
 runtime-benchmarks = [
 	"frame-benchmarking",
+	"pallet-permissions/runtime-benchmarks",
 	"schnorrkel",
 ]
